deal cut off template first page

+++ Libraries/xDC.Domain/ISSD_TS/TS_TotalFlow.cs
@@ -10,7 +10,7 @@ namespace xDC.Domain.ISSD_TS
     {
         public DateTime SettlementDate { get; set; }
         public string Currency { get; set; }
-        public decimal Inflow { get; set; }
-        public decimal Outflow { get; set; }
+        public double Inflow { get; set; }
+        public double Outflow { get; set; }
     }
 }
